fix: arrumo erros nos models, sobre o db

+++ src/main/java/com/jpoltramari/library_api/domain/model/Permissions.java
@@ -15,7 +15,7 @@ public class Permissions {
     @EqualsAndHashCode.Include
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+    private Long id;
 
     @Column(nullable = false, unique = true, length = 50)
     private String name;
